requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.3
easydict==1.13
faiss-cpu==1.12.0
fonttools==4.59.1
idna==3.10
imageio==2.37.0
insightface==0.7.3
joblib==1.5.1
kiwisolver==1.4.9
lazy_loader==0.4
matplotlib==3.10.5
networkx==3.5
numpy==2.3.2
onnx==1.18.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.3.2
pillow==11.3.0
prettytable==3.16.0
protobuf==6.32.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.15.3
simsimd==6.5.1
six==1.17.0
stringzilla==3.12.6
threadpoolctl==3.6.0
tifffile==2025.6.11
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
